Color and Hackity hack hack
Added linux color codes. It was nice to have a splash of color here and there!
Also I added a return code backdoor so that this program can be used as a sort of "dll" of sorts. You can call system("./SM64FM -yop 0 0 0 0 sav.eep out.eep") to get the value of Bomb-omb battle feild star one for example. -yop is the trigger word to activate this back door